Restructured test content models
