+++ examples/RazorWeb/Models/Content/SpecialProduct.cs
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) .NET Foundation and Contributors
+ *
+ * This software may be modified and distributed under the terms
+ * of the MIT license.  See the LICENSE file for details.
+ *
+ * http://github.com/tidyui/coreweb
+ *
+ */
+
+using Piranha.AttributeBuilder;
+using Piranha.Models;
+
+namespace RazorWeb.Models
+{
+    [ContentType(Title = "Special Product")]
+    public class SpecialProduct : Product<SpecialProduct>
+    {
+    }
+}
